requirements.txt
setuptools==28.8.0
pytest==3.9.3
pymongo==3.7.2
scipy==1.2.0
torchvision==0.2.0
jsonpickle==0.9.6
requests==2.18.4
ipdb==0.11
tqdm==4.28.1
torch==1.0.1.post2
pycocotools==2.0.0
munch==2.3.2
objgraph==3.4.0
docopt==0.6.2
mock==2.0.0
imgaug==0.2.8
pandas==0.22.0
opencv_contrib_python==3.4.0.12
matplotlib==2.1.0
tensorflow_gpu==1.12.0
colorama==0.4.0
wrapt==1.10.11
numpy==1.16.2
h5py==2.9.0
py_cpuinfo==4.0.0
Pillow==5.4.1
GitPython==2.1.11
hashfs==0.7.1
Mako==1.0.7
sphinx_rtd_theme==0.4.3
SQLAlchemy==1.3.1
telegram==0.0.1
tensorboardX==1.6
tensorflow==1.13.1
tinydb==3.13.0
tinydb_serialization==1.0.4
